requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
dj-database-url==2.2.0
Django==4.2
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-crispy-forms==2.0
django-embed-video==1.4.9
django-js-asset==2.2.0
django-parler==2.3
django_csp==3.8
fontawesomefree==6.5.1
idna==3.7
pillow==10.3.0
psycopg2-binary==2.9.9
pytz==2024.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
